update holiday-data fetch method
Data source is changed to the Open Data API of https://data.taipei . Because the api has a limit of 1000 result, the data will be capped at 2021/4/23. So a change need to made to parallelly fetch different segments of data and merge them to retrieve the complete holiday datas.

For now, it only retrieves the first 2000 data. The api has only 1077 results, so I figure 2000 may be enough until a more refined solution is made.